refactor to bis_graph, mostly working
