--- .gitignore
@@ -1,2 +1,0 @@
-/osrs-wiki-scraper/img
-/osrs-wiki-scraper/graphs
