+++ osrs-wiki-scraper/bis_graph/.gitignore
@@ -1,0 +1,1 @@
+/graphs
